Split ChordDiagram into separate module
